+++ modules/control/units/controlnet.py
@@ -382,22 +382,6 @@ class ControlNet():
                         self.model = sdnq_quantize_model(self.model)
                     except Exception as e:
                         log.error(f'Control {what} model SDNQ Compression failed: id="{model_id}" {e}')
-                elif "Control" in opts.optimum_quanto_weights:
-                    try:
-                        log.debug(f'Control {what} model Optimum Quanto: id="{model_id}"')
-                        model_quant.load_quanto('Load model: type=Control')
-                        from modules.model_quant import optimum_quanto_model
-                        self.model = optimum_quanto_model(self.model)
-                    except Exception as e:
-                        log.error(f'Control {what} model Optimum Quanto: id="{model_id}" {e}')
-                elif "Control" in opts.torchao_quantization:
-                    try:
-                        log.debug(f'Control {what} model Torch AO: id="{model_id}"')
-                        model_quant.load_torchao('Load model: type=Control')
-                        from modules.model_quant import torchao_quantization
-                        self.model = torchao_quantization(self.model)
-                    except Exception as e:
-                        log.error(f'Control {what} model Torch AO: id="{model_id}" {e}')
                 if self.device is not None:
                     sd_models.move_model(self.model, self.device)
                 if "Control" in opts.cuda_compile: